[plugin.video.viaplay] 1.0.8
