+++ plugin.video.viaplay/addon.xml
@@ -1,13 +1,12 @@
 ﻿<?xml version="1.0" encoding="UTF-8"?>
 <addon id="plugin.video.viaplay"
-       version="1.0.7.1"
+       version="1.0.8"
        name="Viaplay"
        provider-name="emilsvennesson">
   <requires>
     <import addon="xbmc.python" version="2.25.0"/>
     <import addon="script.module.requests" version="2.9.1"/>
     <import addon="script.module.iso8601" version="0.1.11"/>
-    <import addon="script.module.m3u8" version="0.2.10"/>
   </requires>
   <extension point="xbmc.python.pluginsource" library="default.py">
     <provides>video</provides>
@@ -15,16 +14,16 @@
   <extension point="xbmc.addon.metadata">
     <description lang="en">Watch content from Viaplay.</description>
     <description lang="sv">Titta på innehåll från Viaplay.</description>
-    <news>2017.01.25 v1.0.7.1[CR]+ Update for Krypton[CR]+ Fix stream URL retrieval</news>
+    <news>2017.04.10 v1.0.8[CR]+ Fix audio on newer Viaplay content[CR]+ Fix icon/fanart[CR]+ Move addon code to addon.py</news>
     <platform>all</platform>
     <language>sv dk no fi en</language>
     <license>GNU GENERAL PUBLIC LICENSE. Version 3, 29 June 2007</license>
     <source>https://github.com/emilsvennesson/kodi-viaplay</source>
     <forum>http://forum.kodi.tv/showthread.php?tid=286387</forum>
-    <disclaimer>This addon requires you to have a subscription to Viaplay.[CR]This addon is completely unofficial and is not endorsed by Viaplay in any way.</disclaimer>
+    <disclaimer>This add-on requires you to have a subscription to Viaplay.[CR]This add-on is completely unofficial and is not endorsed by Viaplay in any way.</disclaimer>
+    <assets>
+      <icon>resources/art/icon.png</icon>
+      <fanart>resources/art/fanart.jpg</fanart>
+    </assets>
   </extension>
-  <assets>
-    <icon>resources/art/icon.png</icon>
-    <fanart>resources/art/fanart.jpg</fanart>
-  </assets>
 </addon>
