+++ plugin.video.viaplay/changelog.txt
@@ -1,4 +1,9 @@
-﻿2017.01.25 v1.0.7.1
+﻿2017.04.10 v1.0.8
++ Fix audio on newer Viaplay content
++ Fix icon/fanart
++ Move addon code to addon.py
+
+2017.01.25 v1.0.7.1
 + Additional stream URL fix
 
 2017.01.25 v1.0.7
